requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via black, virtualenv
asgiref==3.2.7            # via django
assertpy==1.0             # via -r requirements.in
attrs==19.3.0             # via black, pytest
black==19.10b0            # via -r requirements.in
boto3==1.12.26            # via -r requirements.in
botocore==1.15.31         # via boto3, s3transfer
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via black
distlib==0.3.0            # via virtualenv
django-anymail==7.0.0     # via -r requirements.in
django-cors-headers==3.2.1  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django==3.0.4             # via -r requirements.in, django-anymail, django-cors-headers, django-filter, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
docutils==0.15.2          # via botocore
envparse==0.2.0           # via -r requirements.in
factory-boy==2.12.0       # via -r requirements.in
faker==4.0.2              # via factory-boy
filelock==3.0.12          # via tox, virtualenv
idna==2.9                 # via requests
invoke==1.4.1             # via -r requirements.in
isort==4.3.21             # via -r requirements.in
jmespath==0.9.5           # via boto3, botocore
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest, tox
pathspec==0.7.0           # via black
pluggy==0.13.1            # via pytest, tox
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.8.1                 # via pytest, tox
pyparsing==2.4.6          # via packaging
pytest-django==3.8.0      # via -r requirements.in
pytest==5.4.1             # via -r requirements.in, pytest-django
python-dateutil==2.8.1    # via botocore, faker
pytz==2019.3              # via django
regex==2020.2.20          # via black
requests==2.23.0          # via django-anymail
s3transfer==0.3.3         # via boto3
six==1.14.0               # via django-anymail, packaging, python-dateutil, tox, virtualenv
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.0              # via black, tox
tox==3.14.5               # via -r requirements.in
typed-ast==1.4.1          # via black
urllib3==1.25.8           # via botocore, requests
uwsgi==2.0.18             # via -r requirements.in
virtualenv==20.0.14       # via tox
wcwidth==0.1.9            # via pytest
